refactor: move all pages to top level, rather than nested inside home
