+++ site/src/jsMain/kotlin/web/android/dev/pe/pages/resources/RecommendedResources.kt
@@ -1,10 +1,10 @@
-package web.android.dev.pe.pages.home.pages.resources
+package web.android.dev.pe.pages.resources
 
 import androidx.compose.runtime.Composable
 import com.varabyte.kobweb.core.Page
 import web.android.dev.pe.components.MainSite
-import web.android.dev.pe.pages.home.pages.resources.sections.BlogPostsSection
-import web.android.dev.pe.pages.home.pages.resources.sections.PlaylistsSection
+import web.android.dev.pe.pages.resources.sections.BlogPostsSection
+import web.android.dev.pe.pages.resources.sections.PlaylistsSection
 
 @Page("/resources")
 @Composable
